+++ modin/experimental/pandas/__init__.py
@@ -1,0 +1,7 @@
+from modin.pandas import *
+from .io_exp import read_sql
+import warnings
+
+warnings.warn("\nThank you for the Modin Experimental pandas API. "
+              "\nPlease note that some of these APIs deviate from pandas in order to "
+              "provide improved performance.")
